WIP fixing tests
